define coumn with defult constructor
